# artifacts/api-server/src/routes/orders.ts
import { Router } from "express";
import { OpenOrderBody, CloseOrderResponse } from "@workspace/api-zod";
import { RISK_PROFILES, BASE_PRICES } from "../lib/demo-data";

const router = Router();

const DEMO_MODE = () => !process.env.BYBIT_API_KEY || !process.env.BYBIT_API_SECRET;

router.post("/orders/open", async (req, res): Promise<void> => {
  const parsed = OpenOrderBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { symbol, side, riskMode } = parsed.data;
  const profile = RISK_PROFILES[riskMode];

  if (DEMO_MODE()) {
    const price = BASE_PRICES[symbol] ?? 100;
    const orderId = `DEMO-${Date.now()}`;
    req.log.info({ symbol, side, riskMode, leverage: profile.leverage }, "Demo order opened");
    res.status(201).json({
      success: true,
      message: `[DEMO] ${side === "Buy" ? "LONG" : "SHORT"} abierto en ${symbol} @ $${price.toFixed(2)} con ${profile.leverage}x — Capital: ${profile.capitalPercent}%`,
      orderId,
    });
    return;
  }

  // TODO: Execute real order on Bybit Unified Trading API
  res.status(201).json({
    success: true,
    message: "Orden enviada a Bybit",
    orderId: `BYBIT-${Date.now()}`,
  });
});

router.post("/orders/close", async (req, res): Promise<void> => {
  const { symbol } = req.body as { symbol: string };
  if (!symbol) {
    res.status(400).json({ error: "symbol required" });
    return;
  }

  if (DEMO_MODE()) {
    req.log.info({ symbol }, "Demo order closed");
    res.json(CloseOrderResponse.parse({
      success: true,
      message: `[DEMO] Posición de ${symbol} cerrada exitosamente`,
      orderId: `CLOSE-${Date.now()}`,
    }));
    return;
  }

  res.json(CloseOrderResponse.parse({
    success: true,
    message: `Posición de ${symbol} cerrada en Bybit`,
    orderId: `CLOSE-${Date.now()}`,
  }));
});

export default router;