# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Crypto Trading Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid password
...
  /market/ticker:
    get:
      operationId: getTicker
      tags: [market]
      summary: Get live ticker for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticker data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticker"
  /market/candles:
    get:
      operationId: getCandles
      tags: [market]
      summary: Get candlestick data
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: interval
          in: query
          required: true
          schema:
            type: string
            enum: [1, 5, 15, 60, 240]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Candle data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
  /market/funding-rate:
    get:
      operationId: getFundingRate
      tags: [market]
      summary: Get funding rate for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Funding rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundingRate"
  /market/open-interest:
    get:
      operationId: getOpenInterest
      tags: [market]
      summary: Get open interest
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Open interest data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenInterest"
...
  /orders/open:
    post:
      operationId: openOrder
      tags: [orders]
      summary: Open a position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenOrderBody"
      responses:
        "201":
          description: Order opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"
  /orders/close:
    post:
      operationId: closeOrder
      tags: [orders]
      summary: Close a position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseOrderBody"
      responses:
        "200":
          description: Order closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"
...
components:
  schemas:
...
    LoginBody:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AuthStatus:
      type: object
      required: [authenticated, demoMode]
      properties:
        authenticated:
          type: boolean
        demoMode:
          type: boolean
    Ticker:
      type: object
      required: [symbol, price, change24h, changePercent24h, high24h, low24h, volume24h, markPrice, indexPrice]
      properties:
        symbol:
          type: string
        price:
          type: number
        change24h:
          type: number
        changePercent24h:
          type: number
        high24h:
          type: number
        low24h:
          type: number
        volume24h:
          type: number
        markPrice:
          type: number
        indexPrice:
          type: number
    Candle:
      type: object
      required: [time, open, high, low, close, volume]
      properties:
        time:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
    FundingRate:
      type: object
      required: [symbol, rate, nextFundingTime, predictedRate]
      properties:
        symbol:
          type: string
        rate:
          type: number
        nextFundingTime:
          type: number
        predictedRate:
          type: number
    OpenInterest:
      type: object
      required: [symbol, openInterest, openInterestValue]
      properties:
        symbol:
          type: string
        openInterest:
          type: number
        openInterestValue:
          type: number
...
    OpenOrderBody:
      type: object
      required: [symbol, side, riskMode]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [Buy, Sell]
        riskMode:
          type: string
          enum: [conservative, risky, ultra]
        leverage:
          type: ["integer", "null"]
        capitalPercent:
          type: ["number", "null"]
    CloseOrderBody:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
        side:
          type: ["string", "null"]
    OrderResult:
      type: object
      required: [success, message, orderId]
      properties:
        success:
          type: boolean
        message:
          type: string
        orderId:
          type: string
...